+++ sample/src/com/viewpagerindicator/sample/SampleCirclesInitialPage.java
@@ -3,7 +3,7 @@ package com.viewpagerindicator.sample;
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
 import com.viewpagerindicator.CirclePageIndicator;
-import com.viewpagerindicator.R;
+import com.viewpagerindicator.sample.R;
 
 public class SampleCirclesInitialPage extends BaseSampleActivity {
 	@Override
@@ -16,9 +16,9 @@ public class SampleCirclesInitialPage extends BaseSampleActivity {
 		mPager = (ViewPager)findViewById(R.id.pager);
 		mPager.setAdapter(mAdapter);
 		
-		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
-		indicator.setViewPager(mPager);
-		indicator.setCurrentItem(mAdapter.getCount() - 1);
+		mIndicator = (CirclePageIndicator)findViewById(R.id.indicator);
+		mIndicator.setViewPager(mPager);
+		mIndicator.setCurrentItem(mAdapter.getCount() - 1);
 		
 		//You can also do: indicator.setViewPager(pager, initialPage);
 	}