+++ sample/src/com/viewpagerindicator/sample/SampleTitlesCenterClickListener.java
@@ -2,12 +2,14 @@ package com.viewpagerindicator.sample;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
+import android.widget.Toast;
 
-import com.viewpagerindicator.R;
+import com.viewpagerindicator.sample.R;
 import com.viewpagerindicator.TitlePageIndicator;
 import com.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
+import com.viewpagerindicator.TitlePageIndicator.OnCenterItemClickListener;
 
-public class SampleTitlesUnderline extends BaseSampleActivity {
+public class SampleTitlesCenterClickListener extends BaseSampleActivity implements OnCenterItemClickListener {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -21,5 +23,12 @@ public class SampleTitlesUnderline extends BaseSampleActivity {
 		TitlePageIndicator indicator = (TitlePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
 		indicator.setFooterIndicatorStyle(IndicatorStyle.Underline);
+		indicator.setOnCenterItemClickListener(this);
+        mIndicator = indicator;
 	}
+
+    @Override
+    public void onCenterItemClick(int position) {
+        Toast.makeText(this, "You clicked the center title!", Toast.LENGTH_SHORT).show();
+    }
 }