+++ sample/src/com/viewpagerindicator/sample/SampleTitlesInitialPage.java
@@ -3,7 +3,7 @@ package com.viewpagerindicator.sample;
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
 
-import com.viewpagerindicator.R;
+import com.viewpagerindicator.sample.R;
 import com.viewpagerindicator.TitlePageIndicator;
 
 public class SampleTitlesInitialPage extends BaseSampleActivity {
@@ -17,9 +17,9 @@ public class SampleTitlesInitialPage extends BaseSampleActivity {
 		mPager = (ViewPager)findViewById(R.id.pager);
 		mPager.setAdapter(mAdapter);
 		
-		TitlePageIndicator indicator = (TitlePageIndicator)findViewById(R.id.indicator);
-		indicator.setViewPager(mPager);
-		indicator.setCurrentItem(mAdapter.getCount() - 1);
+		mIndicator = (TitlePageIndicator)findViewById(R.id.indicator);
+		mIndicator.setViewPager(mPager);
+		mIndicator.setCurrentItem(mAdapter.getCount() - 1);
 		
 		//You can also do: indicator.setViewPager(pager, initialPage);
 	}