+++ sample/src/com/viewpagerindicator/sample/SampleTitlesStyledMethods.java
@@ -3,7 +3,7 @@ package com.viewpagerindicator.sample;
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
 
-import com.viewpagerindicator.R;
+import com.viewpagerindicator.sample.R;
 import com.viewpagerindicator.TitlePageIndicator;
 import com.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
 
@@ -19,15 +19,17 @@ public class SampleTitlesStyledMethods extends BaseSampleActivity {
 		mPager.setAdapter(mAdapter);
 		
 		TitlePageIndicator indicator = (TitlePageIndicator)findViewById(R.id.indicator);
+		mIndicator = indicator;
 		indicator.setViewPager(mPager);
 		
 		final float density = getResources().getDisplayMetrics().density;
-		indicator.setBackgroundColor(0xFFDDDDDD);
-		indicator.setFooterColor(0xFFEE3333);
-		indicator.setFooterLineHeight(2 * density);
-		indicator.setFooterIndicatorStyle(IndicatorStyle.None);
-		indicator.setTextColor(0xFF999999);
+		indicator.setBackgroundColor(0x18FF0000);
+		indicator.setFooterColor(0xFFAA2222);
+		indicator.setFooterLineHeight(1 * density); //1dp
+		indicator.setFooterIndicatorHeight(3 * density); //3dp
+		indicator.setFooterIndicatorStyle(IndicatorStyle.Underline);
+		indicator.setTextColor(0xAA000000);
 		indicator.setSelectedColor(0xFF000000);
-		indicator.setSelectedBold(false);
+		indicator.setSelectedBold(true);
 	}
 }