+++ sample/src/com/viewpagerindicator/sample/SampleTitlesTriangle.java
@@ -5,13 +5,12 @@ import android.support.v4.view.ViewPager;
 
 import com.viewpagerindicator.sample.R;
 import com.viewpagerindicator.TitlePageIndicator;
+import com.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
 
-public class SampleTitlesStyledTheme extends BaseSampleActivity {
+public class SampleTitlesTriangle extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		
-		//The look of this sample is set via a style in the manifest
 		setContentView(R.layout.simple_titles);
 		
 		mAdapter = new TestTitleFragmentAdapter(getSupportFragmentManager());
@@ -19,7 +18,9 @@ public class SampleTitlesStyledTheme extends BaseSampleActivity {
 		mPager = (ViewPager)findViewById(R.id.pager);
 		mPager.setAdapter(mAdapter);
 		
-		mIndicator = (TitlePageIndicator)findViewById(R.id.indicator);
-		mIndicator.setViewPager(mPager);
+		TitlePageIndicator indicator = (TitlePageIndicator)findViewById(R.id.indicator);
+		indicator.setViewPager(mPager);
+		indicator.setFooterIndicatorStyle(IndicatorStyle.Triangle);
+        mIndicator = indicator;
 	}
 }